refactor: 불필요 코드 제거 및 코드 리팩터링
